Started moving towards embedding the ap. Added support functions for navigation. Included trig functions for nav. Extensively modified the ap and nav code to protect against zero division and domain errors

+++ InnerOuterLoop/innerOuterInit.m
@@ -3,23 +3,23 @@ derivativesConstant = 5;
 apSampleTime = 0.01;
 
 %% Control Surface Limits
-deLimit = 15*pi/180;
-daLimit = 25*pi/180;
-dtLimit = 1;
-drLimit = 10*pi/180;
+deLimit = single(15*pi/180);
+daLimit = single(25*pi/180);
+dtLimit = single(1);
+drLimit = single(10*pi/180);
 
 %% Integrator Limits
-uHoldLimit = 0.85*dtLimit;          % Controls throttle
-heightErrorLimit = 15*pi/180;       % this loop generates theta_command
-pitchErrorLimit = 0.9*deLimit;      % controls elevator
-yawDamperLimit = 0.9*drLimit;       % controls rudder
-rollControlLimit = 0.95*daLimit;     % controls aileron
+uHoldLimit = single(0.85*dtLimit);          % Controls throttle
+heightErrorLimit = single(15*pi/180);       % this loop generates theta_command
+pitchErrorLimit = single(0.9*deLimit);      % controls elevator
+yawDamperLimit = single(0.9*drLimit);       % controls rudder
+rollControlLimit = single(0.95*daLimit);     % controls aileron
 
 %% Command Limits
-pitchCommandLimit = 15*pi/180;
-bankCommandLimit = 40*pi/180;
+pitchCommandLimit = single(15*pi/180);
+bankCommandLimit = single(40*pi/180);
 
 %% Nav Values
-R = 200;
-L1_base = 200;
-U_comm = 25;
+R = single(200);
+L1_base = single(200);
+U_comm = single(25);
